pyproject.toml
[tool.poetry]
name = "aws_chatbot"
version = "0.1.0"
description = ""
authors = ["Michael Buono <michaelbuono5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.1"
langchain-openai = "^0.1.4"
ruff = "^0.4.2"
pytest = "^8.2.0"
langchain-core = "^0.1.47"
langchain = "^0.1.16"
ipykernel = "^6.29.4"
python-dotenv = "^1.0.1"
boto3 = "^1.34.95"
isort = "^5.13.2"
uuid = "^1.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
